ai-services/anomaly-sage/requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
python-dotenv
httpx
SQLAlchemy
PyJWT
python-jose[cryptography]
opentelemetry-instrumentation-sqlalchemy
numpy